refactored the SummarySourceSinkManager to use an external re-usable resolver
